Extra read receipt logic structure to an abstract class
